Refact : restructure common component
