+++ src/component/common/modal/Alert.js
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 
-import PillShadowButton from "../PillShadowButton";
+import PillShadowButton from "../button/PillShadowButton";
 
 function Alert(props) {
     const onFirstClick = () => {
