+++ src/component/common/modal/DetailPopup.js
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 import { useState, forwardRef, useImperativeHandle } from 'react';
 
-import EmotionTag from './EmotionTag';
+import EmotionTag from '../widget/EmotionTag';
 import { BsChevronCompactLeft, BsChevronCompactRight } from 'react-icons/bs';
 
 function DetailPopup(props, ref) {
