+++ src/component/common/widget/EmotionTag.js
