+++ src/component/pages/ErrorPage/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 
 import { Container } from "./style";
-import Header from "../../common/Header";
+import Header from "../../common/widget/Header";
 
 function ErrorPage(props) {
     return (
