+++ src/component/pages/FeedPage/index.js
@@ -1,9 +1,9 @@
 import React, { useState, useEffect } from "react";
 
 import { Container, PostList, Info, WBR } from "./style";
-import Header from "../../common/Header";
-import PostCard from "../../common/PostCard";
-import FloatingButton from "../../common/FloatingButton";
+import Header from "../../common/widget/Header";
+import PostCard from "../../common/card/PostCard";
+import FloatingButton from "../../common/button/FloatingButton";
 import { GiPencil } from "react-icons/gi";
 import { MdKeyboardArrowUp } from "react-icons/md"
 import Progress from "../../common/modal/Progress";
