+++ src/component/pages/LandingPage/index.js
@@ -1,8 +1,8 @@
 import React from "react";
 
 import { Container, Title, Text, Logo, Wrap, ImagePart, TextPart, MainPart, CharacterImg, FeedImg, RecommendImg, Footer, FooterContent } from "./style";
-import Header from "../../common/Header";
-import PillButton from "../../common/PillButton";
+import Header from "../../common/widget/Header";
+import PillButton from "../../common/button/PillButton";
 import logo from "../../../image/logo_img.svg";
 import character from "../../../image/LandingPageImg/character.png";
 import feed from "../../../image/LandingPageImg/feed.png";
