+++ src/component/pages/WritePage/index.js
@@ -2,10 +2,10 @@ import React, { useState } from "react";
 
 import emotions from "../../../utils/Emotions";
 import { Container, Body, TagContainer, Tag, TextSection, TextArea, CheckBoxContainer, ButtonSection } from "./style";
-import EmotionTag from "../../common/EmotionTag";
-import Header from "../../common/Header";
-import PillButton from "../../common/PillButton";
-import CheckBox from "../../common/CheckBox";
+import EmotionTag from "../../common/widget/EmotionTag";
+import Header from "../../common/widget/Header";
+import PillButton from "../../common/button/PillButton";
+import CheckBox from "../../common/input/CheckBox";
 import Alert from "../../common/modal/Alert";
 
 import * as api from "../../../utils/api";
